requirements.txt
streamlit
pytesseract
pdf2image
python-docx
Pillow
PyMuPDF
